extract Login page as common component
