+++ packages/applications/src/App.config.json
@@ -1,6 +1,7 @@
 {
   "appTitle": "Kernel Application",
   "logoUrl": "/icon-192x192.png",
+  "homeUrl": "/apply",
   "minRole": 1000,
   "navbar": {
     "links": []
