+++ packages/common/src/assets/images/admin_bg.png
