+++ packages/meet/src/App.js
@@ -8,11 +8,10 @@
 
 import React, { Suspense, lazy } from 'react'
 import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom'
-import { ServicesProvider, Loading } from '@kernel/common'
+import { ServicesProvider, Loading, Login } from '@kernel/common'
 
 import 'App.css'
 
-const Login = lazy(() => import('views/Login.js'))
 const Create = lazy(() => import('views/Create.js'))
 const Edit = lazy(() => import('views/Edit.js'))
 const Browse = lazy(() => import('views/Browse.js'))
