+++ packages/profile/src/App.js
@@ -8,11 +8,10 @@
 
 import React, { Suspense, lazy } from 'react'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import { ServicesProvider, Loading } from '@kernel/common'
+import { ServicesProvider, Loading, Login } from '@kernel/common'
 
 import 'App.css'
 
-const Login = lazy(() => import('views/Login'))
 const Profile = lazy(() => import('views/Profile'))
 
 const App = () => {
